Removing unnecessary code and logic
